Refactor image builds
This change introduces multi-stage image builds where it uses
the upstream Golang image to build the project and then places
it in an empty, scratch container to run it inside.

This increases the security of the container significantly due
to the fact that you don't have an operating system alongside
it that you need to keep up to date.  It also drives the size
of the image down.

This does introduce a small behaviour change in that you must
manually setup your arguments and you no longer automatically
have the configuration file auto-detected, but it's probably
better for users to explicitly configure things rather than
have it implicitly happen.