wlansta: Record wifi signal strength
signal-strength-ch0 is only accurate for devices with one antenna. Many
mikrotik devices have two, or even three channels, and just looking at ch0 is
not helpful in those cases.

Instead, look at the property 'signal-strength'. Annoyingly, it is does come as
a string with the negotiated bandwidth attached (e.g.,
signal-strength=-66dBm@HT20-7) which needs to be parsed out.