Support wireless metrics for `wifiwave2` devices (#155)
from https://forum.mikrotik.com/viewtopic.php?t=195124#p999722: wifiwave2
is an implementation of drivers from the manufacturer of the chipset, rather
than an in-house written driver (which wireless is). So there are many small
details that are missing or incomplete...

wifiwave2 has a slightly different API and has fewer properties to query.
This means that not all metrics are available for wifiwave2 devices. Also,
the implementation aims at keeping the name metric names for wifiwave2
devices to ensure that existing dashboard continue to work without any
modification.